[LIS-17] Renamed tests. Changed structure of tests.
