+++ Test/Unit/Facade/Handlers/SkipItem/SkippedItemFixerDataProvider.php
@@ -6,7 +6,7 @@
  * @package Mygento_Base
  */
 
-namespace Mygento\Base\Test\Unit\Facade\SkipItem;
+namespace Mygento\Base\Test\Unit\Facade\Handlers\SkipItem;
 
 use Mygento\Base\Test\OrderMockBuilder;
